Rename directories to follow naming convention.
