+++ app/src/main/java/com/example/weatherapp11/models/WeatherInfo.kt
@@ -1,4 +1,4 @@
-package com.example.weatherapp11.Model
+package com.example.weatherapp11.models
 
 import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
